Improve purge nano script

+++ purge-nano.sh
@@ -1,3 +1,8 @@
-apt-get purge -y nano
-rm /etc/alternatives/editor
-ln -s /usr/bin/vim /etc/alternatives/editor
+#
+# Set vim as default editor and purge nano
+#
+if dpkg -l | grep nano >/dev/null;then
+    apt-get purge -y nano
+    rm /etc/alternatives/editor
+    ln -s /usr/bin/vim /etc/alternatives/editor
+fi
